@@ -5,8 +5,8 @@ configurations {
 }
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 
     packagingOptions {
         exclude('org/apache/commons/codec/language/bm/*.txt') // 124 files (uncompressed 215kb) with unused language specific rules.
@@ -14,9 +14,9 @@ android {
     defaultConfig {
         applicationId 'fr.skyost.timetable'
         minSdkVersion 17
-        targetSdkVersion 27
-        versionCode 46
-        versionName 'v2.0.9'
+        targetSdkVersion 28
+        versionCode 49
+        versionName 'v2.1'
         vectorDrawables.useSupportLibrary = true
         javaCompileOptions {
             annotationProcessorOptions {
@@ -43,21 +43,17 @@ android {
 }
 
 dependencies {
-    def supportVersion = '27.1.1'
-    def lifecycleVersion = '1.1.1'
-    def roomVersion = '1.1.1'
-
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
-    implementation "com.android.support:appcompat-v7:$supportVersion"
-    implementation "com.android.support:design:$supportVersion"
-    implementation "com.android.support:support-v4:$supportVersion"
+    implementation 'androidx.appcompat:appcompat:1.1.0-alpha01'
+    implementation 'com.google.android.material:material:1.1.0-alpha01'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
 
-    implementation "android.arch.lifecycle:extensions:$lifecycleVersion"
-    implementation "android.arch.lifecycle:common-java8:$lifecycleVersion"
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.0.0'
+    implementation 'androidx.lifecycle:lifecycle-common-java8:2.0.0'
 
-    implementation "android.arch.persistence.room:runtime:$roomVersion"
-    annotationProcessor "android.arch.persistence.room:compiler:$roomVersion"
+    implementation 'androidx.room:room-runtime:2.1.0-alpha03'
+    annotationProcessor 'androidx.room:room-compiler:2.1.0-alpha03'
 
     implementation 'net.sf.biweekly:biweekly:0.6.2'
     implementation 'com.github.apl-devs:appintro:v4.2.3'
