@@ -1,7 +1,6 @@
 package fr.skyost.timetable.fragment.intro;
 
-import android.support.v4.app.Fragment;
-
+import androidx.fragment.app.Fragment;
 import fr.skyost.timetable.activity.IntroActivity;
 
 /**
