@@ -9,30 +9,31 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.design.widget.NavigationView;
-import android.support.design.widget.Snackbar;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.TextView;
 
+import com.google.android.material.navigation.NavigationView;
+import com.google.android.material.snackbar.Snackbar;
 import com.kobakei.ratethisapp.RateThisApp;
 
 import org.joda.time.DateTimeConstants;
 import org.joda.time.LocalDate;
 
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.fragment.app.FragmentTransaction;
 import de.mateware.snacky.Snacky;
 import fr.skyost.timetable.R;
 import fr.skyost.timetable.activity.settings.SettingsActivity;
 import fr.skyost.timetable.fragment.day.DayFragment;
 import fr.skyost.timetable.fragment.default_.DefaultFragment;
 import fr.skyost.timetable.receiver.MainActivitySyncReceiver;
+import fr.skyost.timetable.sync.TimetableSyncService;
 import fr.skyost.timetable.utils.Utils;
 
 /**
@@ -345,6 +346,9 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
 	 */
 
 	private void onTimetableFirstLoaded() {
+		// We start the service (if not done before).
+		startService(new Intent(this, TimetableSyncService.class));
+
 		// We set the required views.
 		final Toolbar toolbar = findViewById(R.id.main_toolbar);
 		setSupportActionBar(toolbar);
