@@ -1,9 +1,8 @@
 package fr.skyost.timetable.lesson.database;
 
-import android.arch.persistence.room.Database;
-import android.arch.persistence.room.RoomDatabase;
-import android.arch.persistence.room.TypeConverters;
-
+import androidx.room.Database;
+import androidx.room.RoomDatabase;
+import androidx.room.TypeConverters;
 import fr.skyost.timetable.lesson.Lesson;
 
 /**
