@@ -3,10 +3,6 @@ package fr.skyost.timetable.lesson.database;
 import android.Manifest;
 import android.accounts.Account;
 import android.accounts.AccountManager;
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.OnConflictStrategy;
-import android.arch.persistence.room.Query;
 import android.content.Context;
 
 import org.joda.time.DateTime;
@@ -21,6 +17,10 @@ import java.util.Collection;
 import java.util.HashSet;
 import java.util.List;
 
+import androidx.room.Dao;
+import androidx.room.Insert;
+import androidx.room.OnConflictStrategy;
+import androidx.room.Query;
 import biweekly.Biweekly;
 import biweekly.ICalendar;
 import biweekly.component.VEvent;
